+++ library/training/trainers/__init__.py
@@ -1,0 +1,6 @@
+# Training Trainers Module
+"""Trainer classes for different training types (PEFT, finetune, etc.)"""
+
+from library.training.trainers.peft_trainer import PeftTrainer
+
+__all__ = ["PeftTrainer"]